% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colSds_cpp}
\alias{colSds_cpp}
\title{Efficient Column Standard Deviations}
\usage{
colSds_cpp(x, n_cores = 1L)
}
\arguments{
\item{x}{Numeric Matrix}

\item{n_cores}{Number of cores to use}
}
\description{
Efficient Column Standard Deviations
}
