% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{derivative_cpp}
\alias{derivative_cpp}
\title{Efficient Numerical Derivative for Matrix (Row-wise)}
\usage{
derivative_cpp(x, gridpts, n_cores = 1L)
}
\arguments{
\item{x}{Numeric Matrix (smoothed values)}

\item{gridpts}{Grid points corresponding to columns of x}

\item{n_cores}{Number of cores to use}
}
\description{
Computes the first derivative using centered differences, mirroring sfsmisc::D1tr.
}
