% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{melt_ice_curves_cpp}
\alias{melt_ice_curves_cpp}
\title{Melt Matrix to Long Format Vector}
\usage{
melt_ice_curves_cpp(x, n_cores = 1L)
}
\arguments{
\item{x}{Numeric Matrix}

\item{n_cores}{Number of cores to use}
}
\description{
Efficiently converts a matrix to a long-format vector (row-major order)
for plotting.
}
