% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dice.R
\name{print.dice}
\alias{print.dice}
\title{Print method for \code{dice} objects.}
\usage{
\method{print}{dice}(x, ...)
}
\arguments{
\item{x}{Object of class \code{dice}.}

\item{...}{Ignored for now.}
}
\description{
Prints a summary of a \code{dice} object.
}
