% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ice.R
\name{print.ice}
\alias{print.ice}
\title{Print method for \code{ice} objects.}
\usage{
\method{print}{ice}(x, ...)
}
\arguments{
\item{x}{Object of class \code{ice}.}

\item{...}{Ignored for now.}
}
\description{
Prints a summary of an \code{ice} object.
}
