% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rowCenter_cpp}
\alias{rowCenter_cpp}
\title{Row-wise Centering}
\usage{
rowCenter_cpp(x, n_cores = 1L)
}
\arguments{
\item{x}{Numeric Matrix}

\item{n_cores}{Number of cores to use}
}
\description{
Centers each row of a matrix by subtracting the row mean.
}
