% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sg_smooth_cpp}
\alias{sg_smooth_cpp}
\title{Savitzky-Golay Filter for Matrix (Row-wise)}
\usage{
sg_smooth_cpp(x, window_size, order, deriv, n_cores = 1L)
}
\arguments{
\item{x}{Matrix to smooth row-wise}

\item{window_size}{Size of the filter window (must be odd)}

\item{order}{Polynomial order}

\item{deriv}{Derivative order (0=smooth, 1=first deriv, etc.)}

\item{n_cores}{Number of cores to use}
}
\description{
Smooths each row of a matrix using a Savitzky-Golay filter.
}
