% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ice.R
\name{summary.dice}
\alias{summary.dice}
\title{Summary function for \code{dice} objects.}
\usage{
\method{summary}{dice}(object, ...)
}
\arguments{
\item{object}{Object of class \code{dice}.}

\item{...}{Ignored for now.}
}
\description{
Alias of \code{print} method.
}
