% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dice.R
\name{summary.ice}
\alias{summary.ice}
\title{Summary function for \code{ice} objects.}
\usage{
\method{summary}{ice}(object, ...)
}
\arguments{
\item{object}{Object of class \code{ice}.}

\item{...}{Ignored for now.}
}
\description{
Alias of \code{print} method.
}
