% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{transform_ice_curves_cpp}
\alias{transform_ice_curves_cpp}
\title{Probability Transformation}
\usage{
transform_ice_curves_cpp(x, method, n_cores = 1L)
}
\arguments{
\item{x}{Numeric Matrix (probabilities)}

\item{method}{1 for centered logodds, 2 for probit}

\item{n_cores}{Number of cores to use}
}
\description{
Efficiently applies logodds or probit transformation to a matrix.
}
