% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhiteWine.R
\docType{data}
\name{WhiteWine}
\alias{WhiteWine}
\title{Data concerning white wine.}
\format{
A data frame of 4898 cases on 12 variables.
}
\source{
K Bache and M Lichman. UCI machine learning repository, 2013.
http://archive.ics.uci.edu/ml
}
\usage{
data(WhiteWine)
}
\description{
The WhiteWine data frame has 4898 rows and 12 columns and
concerns white wines from a region in Portugal. The response
variable, quality, is a wine quality metric, taken to be the median
preference score of three blind tasters on a scale of 1-10. The 11 covariates
are physicochemical metrics of wine quality such as citric acid content,
sulphates, etc.
}
\keyword{datasets}
