\name{summary.dice}
\alias{summary.dice}
\title{
Summary function for \code{dice} objects.
}
\description{
Calls \code{print} and prints a brief summary of a \code{dice} object.
}
\usage{
\method{summary}{dice}(object,...)
}

\arguments{
  \item{object}{
Object of class \code{dice}.
}
  \item{...}{
Ignored for now.
}
}

