\name{summary.ice}
\alias{summary.ice}
\title{
Summary function for \code{ice} objects.
}
\description{
Calls \code{print} and prints a brief summary of an \code{ice} object.
}
\usage{
\method{summary}{ice}(object,...)
}
\arguments{
  \item{object}{
Object of class \code{ice}.
}
  \item{...}{
Ignored for now.
}
}

